+++ bpf/insn/samples/bpf_insn_defs.h
@@ -11,6 +11,8 @@ BPF_INSN_SAMPLE(strncmp)
 BPF_INSN_SAMPLE(strnstr)
 BPF_INSN_SAMPLE(strlen)
 BPF_INSN_SAMPLE(strnlen)
+BPF_INSN_SAMPLE(send_signal)
+BPF_INSN_SAMPLE(send_signal_thread)
 /**
  * linux v5.15-12938-ge6f2dd0f8067
  * commit e6f2dd0f8067 ("bpf: Add bpf_loop helper")